Fix Kotlin 2.x and Gradle 9.x compilation issues
- Add ExperimentalForeignApi opt-in for native compilation
- Fix Int->ULong for posix API size parameters (read, write, fread, getcwd)
- Replace deprecated toLowerCase() with lowercase()
- Remove deprecated @ThreadLocal annotations (unnecessary with new memory model)
- Fix typealias import qualifier issue (KT-64431) for ReferencePolicy
- Fix remaining camelCase rename misses (TU, Std, TestLib, ToChecked, TEMPLATE_HEADER)
- Update ksrpc coordinates: com.monkopedia -> com.monkopedia.ksrpc, add ksrpc-sockets
- Fix plugin JVM target to 21 to match Java toolchain
- Add kotlin.native.cacheKind.linuxX64=none for Clikt 5.x linking
- Fix Gradle 9 task dependency validation for sourcesJar
- Fix dependsOn null safety for Task?

+++ krapper_gen/src/nativeMain/kotlin/com/monkopedia/krapper/generator/Parsing.kt
@@ -343,16 +343,16 @@ fun generateIncludes(compiler: String) = memScoped {
     val buffer = alloc<ByteVar> {
         EOF
     }
-    write(process.stdIn(), buffer.ptr, 1)
+    write(process.stdIn(), buffer.ptr, 1.toULong())
     close(process.stdIn())
     process.wait()
     val readBuffer = allocArray<ByteVar>(256)
     var fullString = StringBuilder()
-    var amount = read(process.stdOut(), readBuffer, 255)
+    var amount = read(process.stdOut(), readBuffer, 255.toULong())
     while (amount > 0) {
         readBuffer[amount.toInt()] = 0.toByte()
         fullString.append(readBuffer.toKStringFromUtf8())
-        amount = read(process.stdOut(), readBuffer, 255)
+        amount = read(process.stdOut(), readBuffer, 255.toULong())
     }
     val lines = fullString.split("\n")
     val start = lines.indexOf("#include <...> search starts here:")
